requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi
uvicorn
langchain_community
langchain_ollama
langchain_core
langchain_chroma
langchain
pydantic
chromadb
pypdf
requests
beautifulsoup4
aiohttp
#sentence-transformers
pdfminer.six
python-multipart
typing
uuid
certifi
numpy
scikit-learn
nltk
tiktoken
python-dotenv